changes in services

+++ finance-management-service/src/main/java/com/main/dto/BudgetDto.java
@@ -1,17 +1,17 @@
 package com.main.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import lombok.Data;
 
 @Data
 public class BudgetDto {
-	    private int userId;
-	    private double amount;
-	    private String category;
-	    private LocalDate startDate;
-	    private LocalDate endDate;
-	
-	}
+    private int userId;
+    private BigDecimal amount;
+    private String category;
+    private LocalDate startDate;
+    private LocalDate endDate;
+}
